+++ network/network/src/main/java/bisq/network/p2p/node/transport/socket_channel/SocketChannelFactory.java
@@ -15,16 +15,13 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.network.p2p.node.transport.socketchannel;
+package bisq.network.p2p.node.transport.socket_channel;
 
 import bisq.network.common.Address;
 
 import java.io.IOException;
 import java.nio.channels.SocketChannel;
 
-public class ClearNetSocketChannelFactory implements SocketChannelFactory {
-    @Override
-    public SocketChannel getSocketChannelForAddress(Address address) throws IOException {
-        return SocketChannel.open();
-    }
+public interface SocketChannelFactory {
+    SocketChannel getSocketChannelForAddress(Address address) throws IOException;
 }
